Rewrite Meetup class
This is more JavaScript class idiomatic. It also allows for future
abstractions like removing the hard coded URLs.

The original file used tabs so I kept that even though the .vimrc seems
to contradict this.

This also uses UMD for encapsulation which makes it more universal to
be pulled into other websites.